Crate forest_chain_sync

Re-exports

pub use self::consensus::collect_errs;

pub use self::consensus::Consensus;

Modules

consensus

Structs

BadBlockCache

Thread-safe cache for tracking bad blocks. This cache is checked before validating a block, to ensure no duplicate work.

ChainMuxer

The ChainMuxer handles events from the P2P network and orchestrates the chain synchronization.

SyncConfig

Structure that defines syncing configuration options

SyncState

State of the node’s syncing process. This state is different from the general state of the ChainSync process.

TipsetValidator

Enums

SyncStage

Current state of the ChainSyncer using the ChainExchange protocol.