Trait forest_chain_sync::consensus::Consensus

pub trait Consensus: Scale + Debug + Send + Sync + Unpin + 'static {
    type Error: Debug + Display + Send + Sync;

    fn validate_block<'life0, 'async_trait, DB>(
        &'life0 self,
        state_manager: Arc<StateManager<DB>>,
        block: Arc<Block>
    ) -> Pin<Box<dyn Future<Output = Result<(), NonEmpty<Self::Error>>> + Send + 'async_trait>>
    where
        DB: Blockstore + Store + Clone + Sync + Send + 'static + 'async_trait,
        Self: 'async_trait,
        'life0: 'async_trait;
}

The Consensus trait encapsulates consensus specific rules of validation and block creation. Behind the scenes they can farm out the total ordering of transactions to an arbitrary consensus engine, but in the end they package the transactions into Filecoin compatible blocks.

Not all fields will be made use of, however, so the validation of these blocks at least partially have to be trusted to the Consensus component.

Common rules for message ordering will be followed, and can be validated outside by the host system during chain synchronization.

Required Associated Types

type Error: Debug + Display + Send + Sync

Required Methods

fn validate_block<'life0, 'async_trait, DB>(
&'life0 self,
state_manager: Arc<StateManager<DB>>,
block: Arc<Block>
) -> Pin<Box<dyn Future<Output = Result<(), NonEmpty<Self::Error>>> + Send + 'async_trait>>where
DB: Blockstore + Store + Clone + Sync + Send + 'static + 'async_trait,
Self: 'async_trait,
'life0: 'async_trait,

Perform block validation asynchronously and return all encountered errors if failed.

Being asynchronous gives the method a chance to construct a pipeline of validations, i.e. do some common ones before branching out.