Trait forest_chain_sync::consensus::Proposer

pub trait Proposer {
    // Required method
    fn spawn<'life0, 'async_trait, DB, MP>(
        self,
        state_manager: Arc<StateManager<DB>>,
        mpool: Arc<MP>,
        submitter: SyncGossipSubmitter,
        services: &'life0 mut JoinSet<Result<()>>
    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
       where DB: Blockstore + Store + Clone + Sync + Send + 'static + 'async_trait,
             MP: MessagePoolApi + Sync + Send + 'static + 'async_trait,
             Self: 'async_trait,
             'life0: 'async_trait;
}

The Proposer trait expresses the ability to “mine”, or in more general, to propose blocks to the network according to the rules of the consensus protocol.

It is separate from the Consensus trait because it is only expected to be called once, then left to run in the background and try to publish blocks to the network which may or may not be adopted.

It exists mostly as a way for us to describe what kind of dependencies mining processes are expected to take.

Required Methods

fn spawn<'life0, 'async_trait, DB, MP>( self, state_manager: Arc<StateManager<DB>>, mpool: Arc<MP>, submitter: SyncGossipSubmitter, services: &'life0 mut JoinSet<Result<()>> ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>where DB: Blockstore + Store + Clone + Sync + Send + 'static + 'async_trait, MP: MessagePoolApi + Sync + Send + 'static + 'async_trait, Self: 'async_trait, 'life0: 'async_trait,

Start proposing blocks in the background and never return, unless something went horribly wrong. Broadly, they should select messages from the mempool, come up with a total ordering for them, then create blocks and publish them to the network.

To establish total ordering, the proposer might have to communicate with other peers using custom P2P messages, however that is its own concern, the dependencies to implement a suitable network later must come from somewhere else, because they are not common across all consensus variants.

The method returns a vector of handles so that it can start unspecified number of background tasks, which can all be canceled by the main thread if the application needs to exit. The method is async so that it can use async operations to initialize itself, during which it might encounter some errors.