Trait forest_chain_sync::consensus::MessagePoolApi

pub trait MessagePoolApi {
    fn select_signed<'life0, 'life1, 'life2, 'async_trait, DB>(
        &'life0 self,
        state_manager: &'life1 StateManager<DB>,
        base: &'life2 Tipset
    ) -> Pin<Box<dyn Future<Output = Result<Vec<Cow<'_, SignedMessage>>>> + Send + 'async_trait>>
    where
        DB: Blockstore + Store + Clone + Sync + Send + 'static + 'async_trait,
        Self: 'async_trait,
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait;
}

The MessagePoolApi is the window of consensus to the contents of the MessagePool.

It exists to narrow down the possible operations that a consensus engine can do with the MessagePool to only those that it should reasonably exercise, which are mostly read-only queries to get transactions which can be expected to be put in the next block, based on their account nonce values and the current state.

The MessagePool is still expected to monitor the chain growth and remove messages which were included in blocks on its own.

Required Methods

fn select_signed<'life0, 'life1, 'life2, 'async_trait, DB>(
&'life0 self,
state_manager: &'life1 StateManager<DB>,
base: &'life2 Tipset
) -> Pin<Box<dyn Future<Output = Result<Vec<Cow<'_, SignedMessage>>>> + Send + 'async_trait>>where
DB: Blockstore + Store + Clone + Sync + Send + 'static + 'async_trait,
Self: 'async_trait,
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,

Select the set of suitable signed messages based on a tipset we are about to build the next block on.

The result is a Cow in case the source can avoid cloning messages and just return a reference. They will be sent to the data store for storage, but a reference is enough for that.